--- src/main/java/com/backend/CoApp/models/UserModel.java
@@ -1,19 +1,0 @@
-package com.backend.CoApp.models;
-
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-@Document(collection = "users")
-public class UserModel {
-
-    @Id
-    public String id;
-
-    public String firstName;
-    public String lastName;
-
-    public UserModel(String firstName, String lastName) {
-        this.firstName = firstName;
-        this.lastName = lastName;
-    }
-}
